refactor[#405] : deleteTemporaryMarking 공용 api 를 통해 export
