+++ src/features/follow/api/index.ts
@@ -1,3 +1,4 @@
 export * from "./deleteFollowing";
 export * from "./postFollowing";
 export * from "./deleteFollower";
+export * from "./deleteTemporaryMarking";
