+++ src/mocks/handler.ts
@@ -6,7 +6,7 @@ import {
   LOGIN_END_POINT,
   SIGN_UP_END_POINT,
 } from "@/features/auth/constants";
-import { DeleteTemporaryMarkingRequest } from "@/features/follow/api/deleteTemporaryMarking";
+import { DeleteTemporaryMarkingRequest } from "@/features/follow/api";
 import { MY_MARKING_ENDPOINT } from "@/features/follow/constants";
 import { MARKING_END_POINT } from "@/features/marking/constants";
 import { PostChangeRegionRequest } from "@/features/setting/api";
